Clean up and move `UpdateAssetReferences` listener for consistency with new terms listener.
