+++ tests/Listeners/UpdateAssetReferencesTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Assets;
+namespace Tests\Listeners;
 
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Storage;
